fix: rename package, remove setter
